Synchronised responses in DCNet
Due to the received messages coming from different akka actors the calls
to our DCNet are made from different threads. Due to the instant
messages on a local test setup this allowed threads to access the
responses array simultaniously.

By synchronising this behaviour it is no longer possible for different
threads to access this check and therefore enter the
messageRoundCompleted stage.