Update timeout logic
This change makes the timeout logic store the roundId in which they
started and compare it themselves to the current round id. The future
now returns true if a timeout has occured and false otherwise. This
allows the use of onComplete allowing us to properly start the next
round again.

The Thread.sleep calls are broken up in to chunks of 10ms which seems to
prevent the main thread from locking up.

With these changes I am able to make a complete run of at least 1000
messages with no TIMEOUT when testing the ExperimentObserver with a
clique size of 1 (only the center node that is also running the experiment).